Working around the way celery caches result backend settings.
When celery tasks are imported, configuration is immediately parsed and cached.
Unfortunately, pecan has to load celery config at runtime, so pecancelery
overwrites these values as best it can with a custom loader (at runtime).

Backend settings are also explicitly cached, and pecancelery wasn't overriding
this, so the result was that specifying RESULT_* settings (e.g., backend
settings) in your pecan project's config was meaningless, because the result
store would always fall back to the default (AMQP).

This commit provides a workaround that causes RESULT_* settings to be
respected.